INTERLOK-269 Initial support for xinclude via XincProc INTERLOK-269 #time 2d
Added a custom AdapterRegistry implementation that uses XincProc to
pre-process the URL / String that is passed into createAdapter() before
attempting to unmarshal the adapter.

The xinclude implementation is considered "pretty niche/advanced" so
it is not included as part of the standard nightly download/installer.
You have to "depend on it" and download the artifacts via ivy / maven
yourself.