INTERLOK-269 Added in an XStreamMarshaller that supports xinclude INTERLOK-269 #time 2h
Refactored the package names now that we have 2 xinclude implementations

xinclude.AdapterRegistry is probably redundant, as you will most
likely want to use the xinclude.XStreamMarshaller as the default
unmarshaller/ set it where appropriate in the services (e.g. dynamic
services). But we will leave AdapterRegistry in, just for completeness.

Added tests for each unmarshal method.